Update MenuItems docs to match implementation
There were various menu items with wrong parameter documentation in
them. The documentation that was present was about other variable, which
doesn't exist in those functions anymore.

This commit fixes it by updating the documentation, taking the right
documentation from the base class of MenuItem and pasting it in the
different documentations of the derived objects.